TG-113: table renderized on index. List deleted
