--- src/Components/Admin/Projects/List/list.module.css
@@ -1,7 +1,0 @@
-.container {
-  flex: 1;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  text-align: center;
-}
